Add files via upload
"Repository accompanying the MSc thesis on the 'IMPACT OF IMAGE COMPRESSION AND FORMAT CONVERSION ON FINGERPRINT RECOGNITION'. This project provides implementations and analysis scripts to demonstrate the degradation of fingerprint recognition performance due to image compression and format conversion. It includes experiments on minutiae reduction and its influence on matching scores, offering insights into the security and performance implications for biometric systems."